Exception handler 추가

+++ src/main/java/sparta/team6/momo/controller/RecordController.java
@@ -22,7 +22,7 @@ public class RecordController {
     }
 
     @GetMapping("/list")
-    public ResponseEntity<Object> showRecord(@RequestParam int pageNumber) {
+    public ResponseEntity<Object> showRecord(@RequestParam Long pageNumber) {
         List<ShowRecordResponseDto> dtoList = planService.showRecord(pageNumber);
         return ResponseEntity.ok().body(new PagingDto<>(dtoList));
     }
